backend/academy_bot/requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv==1.0.0
